requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
alembic==1.12.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0; python_version >= '3.7'
bcrypt==4.0.1
black==23.3.0
blinker==1.6.3
cachelib==0.9.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==40.0.2
deprecated==1.2.14
dnspython==2.4.2
email-validator==2.0.0.post2
flask==3.0.0
flask-bcrypt==1.0.1
flask-blueprint==1.3.0
flask-cors==4.0.0
flask-restful==0.3.9
flask-sqlalchemy==3.1.1
frozenlist==1.4.0; python_version >= '3.8'
greenlet==3.0.0
gunicorn==20.1.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
jinja2==3.1.2
limits==3.6.0
mako==1.2.4
markdown-it-py==3.0.0
markupsafe==2.1.3
mdurl==0.1.2
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0
openai==0.28.1
ordered-set==4.1.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pygments==2.16.1
pyjwt==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
rich==13.6.0
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0
sqlalchemy==2.0.21
tomli==2.0.1
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.8.0
urllib3==1.26.17
werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2; python_version >= '3.7'
